backend/requirements.txt
# TuCitaSegura Backend - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Firebase
firebase-admin==6.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Deep Learning (opcional, para features avanzadas)
# tensorflow==2.15.0
# torch==2.1.1

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
# deepface==0.0.79  # Descomentar si quieres face recognition

# NLP
# transformers==4.35.2  # Descomentar para NLP avanzado
# spacy==3.7.2
nltk==3.8.1

# Analytics & Forecasting
# prophet==1.1.5  # Descomentar para forecasting
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Geolocalización
geopy==2.4.1
# googlemaps==4.10.0  # Descomentar cuando tengas API key

# Pagos
stripe==7.4.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Background Tasks
celery==5.3.4
redis==5.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
sentry-sdk[fastapi]==1.39.1
